fix(order/upload-payment): TypeError int-string on rate limit check
PHP 8.4 strict arithmetic: session values are stored/retrieved as strings,
so casting with (int) is required before subtracting timestamps.
Also cast the time() call on set() for consistency.

+++ frontend/controllers/OrderController.php
@@ -566,8 +566,8 @@ class OrderController extends Controller
     {
         $session = Yii::$app->session;
         $key = 'upload_attempts_' . $token;
-        $attempts = $session->get($key, 0);
-        $lastAttempt = $session->get($key . '_time', 0);
+        $attempts    = (int)$session->get($key, 0);
+        $lastAttempt = (int)$session->get($key . '_time', 0);
 
         // Сброс счетчика через 15 минут
         if (time() - $lastAttempt > 900) {
@@ -580,7 +580,7 @@ class OrderController extends Controller
         }
 
         $session->set($key, $attempts + 1);
-        $session->set($key . '_time', time());
+        $session->set($key . '_time', (int)time());
     }
 
     /**
